Add files via upload
The given code demonstrates how to swap two integer variables a and b without using a third variable.#include<iostream> it is header file . using namespace std it is template. Int main is main function.The swapping process involves the following steps. a = a + b The new value of a becomes the sum of a and b:a = 8 + 13 = 21 .b = a - b The new value of b is obtained by subtracting the old value of b from the new value of a b = 21 - 13 = 8 then a = a - b The new value of a is obtained by subtracting the new value of b from the current value of a. Return 0 is for successfuly execution of program.