Add files via upload
These is a program of swaping using call by reference. here we use function for call by reference void return type, swap  is function , int is a parameter type and a is parameter name. these before swapping and after swapping thse two type of swapping are before swaping  the no as it is and when we swap it then alter change the number means x,y these are actual number then when we swap it the y,x are swap we get these output